Add in revision summaries to the log_accum.pl file...
I've been using this code in this approximate form for about 6 months now.
I'm keen to hear how useful this is/isn't for people.

It gives an unambiguous indication of how big the commit is, and the
revision numbers are collected into one place to make it easier to find and
follow all the parts of a commit over multiple files.