Fix shar so that it doesn't generate:   # This shar contains:   # length mode       name   # ------ ---------- ------------------------------------------   #   9525 (null) -rw-r--r--   #   3768 (null) -rw-r--r-- 	  ^^^^^^^^^^^ This is a long/off_t printf problem.
The only problem with doing this, is that touching the sources with patch
causes it to want to go and regenerate it's locale databases using some
gnu tools I've never heard of before. :-(  Since patch-ac doesn't actually
change any messages, I've cheated and touched the dependency files in
post-patch.  Please, no pukeing, I know it's a gross hack. ;-)