Update the Modula-3 port with two main goals in mind.
First, change the port so that it builds a much smaller subset of
the SRC distribution.  This eliminates the enormous swap space
requirements of the earlier port, greatly reduces the footprint of
the installed tree, and cuts the size of the package in half.

Second, include many important new patches.  Among them is a slightly
modified version of phkmalloc that is thread-safe for Modula-3.
It eradicates some rare and baffling core dumps that cropped up
from time to time in the previous version of the port.  The Modula-3
runtime itself is careful to use mutual exclusion around calls to
malloc.  But there remained some sneaky backdoor paths into it from
external libraries.

Confession: In the original version of the Modula-3 port, I used
a major version number of 353 for the shared libraries, to correspond
with the SRC version number 3.5.3.  That was a dumb move -- I should
have used 1.  The current update is incompatible at the shared
library level, requiring me to increment the major version number
to 354, even though this is still based on SRC release 3.5.3.  This
is bound to confuse some folks, unfortunately.  I weighed a number
of alternatives, such as (a) cheating and going back to 1, and (b)
using a 4-digit major version such as 3531.  But in the end I
decided that 354 would be the best solution, even though it's
confusing.