Update 1.1.5+patches -> 1.1.6
Note, there are a few important changes to the squid.conf file,
be sure to read the ChangeLog file for the details.  Some names
have been renamed, and a new option needs to be specified if squid
is run as root - it needs to know what uid to change to, it won't run
as root by default.  This may require changes to the port to create a
user or something.