Update exmh-2.0gamma to 2.0delta.  There's lots of bug fixes, as well as things like koi8-r recognition, background news (nntp) processing, ispell support.
Note that there's a horrible hack here to work around a weakness in
bsd.port.mk.  The distributed patches are relative to a subdir of ${WRKSRC}
and bsd.port.mk doesn't seem to allow clean overriding of that to patch.
I appended a second "-d dir" arg to PATCH_DIST_ARGS after the .include
that set up the string with +=.  patch seems to use the last instance of
"-d dir" when there are several online (it's a simple parse-time chdir).

Prompted by: pst (who also sent a partial update)