Wnn ver 6.  Files were already repository copied from japanese/Wnn.
Wnn6 is just a wrapper for the commercial Wnn6 CDROM, so you obviously
need the CDROM to use the port.  Wnn6-lib is the library interface,
and this half is free (this port will be used by the upcoming mule
ports).

PR:		4166
Submitted by:	Satoshi Taoka <taoka@infonets.hiroshima-u.ac.jp>