Import of x11/eterm version 0.6a
Eterm is one of the most configurable terminal emulators available for
the X-Windows system.  It is built as part of the Enlightenment window
manager and as such places the importance of eyecandy and other aesthetic
qualities slightly above stability.  Allows you to use scaled images as your
background for terminal windows, and much more!

PR:		ports/4890 port/4891
Submitted by:	murray@cdrom.com