Update from 1.2.beta11 -> 1.2beta14 Major changes:   - the URL and SHA cache key methods are gone, MD5 is the only option.   - the swap.state system has been redesigned as a transaction log for     much faster restarts and crash recovery.   - SIZE on ftp binary transfers.   - Lots of bug fixes and code cleanups.
From the web page; http://squid.nlanr.net/Squid/1.2.beta/

WARNING, with beta12 we are starting to stuff object metadata into
the beginning of objects on disk.  With beta13 we have changed the
"swap.state" file format to binary instead of text. When you run this
version for the first time you WILL LOSE YOUR CACHE!

The previous warning still stands:
This is BETA software. Do not run this on your production systems.
Logfile formats are subject (and likely) to change at any time without
warning.