Update to 0.9beta1.
This version is now GNU configure driven.

A word of warning for future committers so that you don't fall into
the same trap as I did: The source has an printf-style funtion
"print". This doesn't understand "%u", so the usual print-casting
stuff for long long ints isn't
  printf("%u", (unsigned int)foo);
but
  print("%d", (int)foo);
which obviously looses in some cases, but I didn't take the effort to
extend this function to take 64 bit value (just unsigned is limited as
well).

In this case, it isn't that bad, since it is used to print limits, but
memory limits are internally stored as "mega" or "kilo" anyway.