Add "textproc" as a category.
Actually, can this thing be moved to `textproc' to live with `ispell'?
Having a distributed electronic dictionary in "net" is stupid.  This port is
virtually ispell with /usr/share/dict/words over the network.  Otherwise I vote
to move `xmcd' to `net' since it gets its CD database over the network.