This is the latest version of xemacs - Xemacs 21.1.
This requires xemacs-packages which be will committed next. Xemacs now
allows the user to maintain packages himself. See DESCR.

This supersedes PR 11618, which was based on an older release.

This commit was blasted over a repository copy of xemacs20, which is why
there are so many changes.

Submitted by:	Michael Elbel (me@freebsd.org)