Remove James Raynard from committers:
"I have decided to resign as a committer as it's been a very long time
since I last had the time to work on FreeBSD and things have now got
to the point where not even "speed reading" can help me keep up with
the commit mail."

And thanks to James for being honest about this.  Most folks just
slink away quietly and are never heard from again. :)