Add bsd.gnome.mk that defines gnome-related dependency stuff and hooks in bsd.port.mk to support it.
In particular, bsd.port.mk includes bsd.gnome.mk twice (once in the
pre.mk section, once in the post.mk section) and prints out a message
notifying the user of the existence of WITH_* options when one exists.

It also adds MAKE=${GMAKE} to CONFIGURE_ENV when USE_GMAKE is defined.
(This is not related to gnome, but I assume it's useful somewhere and
it shouldn't hurt.)

Submitted by:	ade, reg