Update to version 2.3.
- Conform to the new port layout, finally.

- Use COMMENT, DESCR, PLIST, PKGINSTALL, PKGDEINSTALL, PKGREQ,
PKGMESSAGE, SCRIPTDIR and PATCHDIR instead of hardcoded file/directory
names. (suggested by will)

- Fix maxchars checker that has been broken.

- Fix "use ldconfig with ||/usr/bin/true" checker. (patch submitted by
sobomax)

- Fix "include the country code in the module alias name" warning.
Besides, "country code" is corrected to "language code".

- Add french and hebrew to the list of lang-specific categories.

- Properly omit the checks against PORTNAME section etc. when the
testee is a slave port.

- Add "INSTALLS_SHLIB may be missing" checker which searches pkg-plist
for `*.so' and `*.so.<nn>'.

- Make it dynamically read bsd.sites.mk so that we no longer need to
keep it always in sync with bsd.sites.mk.  It now should recognize
`/%SUBDIR%/' part too.

- Change `split(/\s+/, "blah blah blah")' to `qw(blah blah blah)'.

- Add some dummy comments that prevent Emacs' CPerl mode from
confusing.

- Add a condition "unless this is a master port" to the warnings that
are specific to master ports, because currently we can't know if a
port is a master port.

Reviewed by:	mharo (MAINTAINER)