Do not use IGNOREFILES.
I'm removing IGNOREFILES because I cannot get a clue about which files are
updated on the webmin site. For example, if any module is updated
twice, it will be listed only once on the site's updates page.

The port will always work because the file's md5 is ignored, but I don't
know when I should bump PORTREVISION. Since the default behavior of the
port is to install the official version without updates, I prefer to have
a md5 mismatch every now and then, so that PORTREVISION can be bumped and
webmin users know when they _actually_ have to upgrade the port.

No user visible changes, so no PORTREVISION bump.