Upgrade this to perl5.6.0 for _knowledgeable_  folks in STABLE who really want it.
This is still marked FORBIDDEN as there may be all sorts of horrible
incompatabilities with the perl5 in "base", but folks who are willing
to override this and experiment are welcome.

Consider this to be only partially supported. I'll happily commit
patches and fixes, but I don't want to answer too many questions.

Sorry!