Fix a bug that prevented updates to be installed.
The line:

  for webmin_module in "${WEBMIN_MODULES}"

must be rewritten as:

  for webmin_module in ${WEBMIN_MODULES}

else the port will not unpack the update tarball.

This is a very stupid bug by me. Pass the pointy hat :-(

PORTREVISION bump: people using the WITH_UPDATES option should upgrade to
this version of the port. People who do not use WITH_UPDATES and run the
stock webmin release, or update it by hand, are unaffected by this patch.