Revived from ports/gcc11. which was lost in the great attic-sweep. Thanks to all the people who had a local copy of it at freefall.
- Udpate to new layout and cleaned with portlint.
- Depends on autconf

This is a 6811 cross compiler for FreeBSD.

The C, C++ compiler is based on the GCC 2.95.2 release. The M68HC1x port is
based on an old port of GCC for 68HC11 made by Otto Lind.  However, it was
intensively modified to generate smaller and faster code. It was enhanced
to support 32-bits and 64-bits integers. The support for IEEE floating point
numbers (32 and 64-bits) is also provided. Source level debugging is available
with DWARF-2 debugging format.