Fix this so it works on FreeBSD/alpha 4.x:
The existing patch does different things for 5.x so as to use the
standard __bswapSIZE(x) inline for 5.x. and adds i386 centric code
for FreeBSD 4.x.  This change to the patch modifies the 4.x code
so as to just call htonX(x) on alpha where __byte_swap_X don't exist.

Approved by: will