Upgrade to 3.4.4 released over a year ago. Obey CFLAGS.
Split the man-pages into the actual texts and MLINKS, of which there are
over a hundred. The vendor used the .so macro, but either it was not
used properly, or there is a bug in our man(1). Either way, MLINKS seems
cleaner.

Move the text in the post-install target into pkg-message and arrange
for showing it after installing the port. The last two items were not
shown to maintainer, but are not expected to raise objections.

TODO -- add WITHOUT_X11 compliance -- a must for some deployments.

Approved by:	demon (maintainer).