# Recorde information which I forgot to recorde in previous commit log.
(1)Update version to 1.38

PR:		ports/48740

and

Submitted by: 	arved,sumikawa

(2)Change the way to fetch to one via cgi.
(The way has a problem. The size of a remote file cannot be obtained.
So the distfile is obtain whenever we do "make fetch".)

(3)mpg123.el supports ogg123. So "ogg123" is added to COMMENT in Makefile

Submitted by:	the author of mpg123.el