Upgrade to version 8.3.5, the latest from ISC. Several bugfixes are included, as well as a few minor enhancements. Check CHANGES for details.
Improve greatly the PORT_REPLACES_BASE option, which now truly replaces
everything, including re-positioning nslookup and nsupdate in sbin, and
writing over the includes, instead of installing them in include/bind.

Unconditionally install dnskeygen in bin, since that's where the base
puts it, and it's only confusing to users if they can't find it after
an installation.

Install isc/ctl.h by hand in the appropriate include directory, since
we install it in the base, and the distribution Makefile does not.