PR:		53013 Submitted by:	Clement Laforet <sheepkiller@cultdeadsheep.org>
Add prelude-nids 0.8.1,
Sensor program of prelude Network Intrusion Detection System

prelude-nids : the Prelude Network Intrusion Detection System is a sensor,
that may be installed on the same machine as the manager or not, which watches
network traffic and looks for familiar patterns. This is functionally
equivalent to Snort (http://www.snort.org).