Remove unneeded RUN_DEPENDS.
Volker mentioned an install with a dependency "postgres" fails if
postgresql is installed with -DWITHOUT_SERVER, though it shouldn't,
and thus suggested a depend-change to "psql", which is also present
for the client-only-install.

The maintainer then submitted the patch which completely removes the
dependency as it "is actually not needed".  I personally am in favor
of the first solution, but Palle is the maintainer.

Reported by:	Volker Stolz <stolz@i2.informatik.rwth-aachen.de>
Submitted by:	Palle Grigensohn <girgen@pingpong.net> (maintainer)
PR:		54252,54287