Update VMware Tools for 4.0.1 build-5289, the latest version of VMware Workstation.
This commit makes vmware-tools4 a slave port of vmware-guestd4, since
both are made from the same tarball, and it would be easy to maintain.
Vmware-guestd4 also installs a kernel module named vmmemctl; you may
want to kldload(8) it with a shell script under rc.d/ directory.

VMware Workstation 4.0.1 does fix "jumping the pointer" bug (yeah, thanks
VMware developer team).  This means that you do not have to install
linux-vmware-toolbox4.  However, it can be used as it should be, I
do not disable and/or remove this port.