- New port: mail/p5-Email-Simple: Simple, fast mail parser
Email::Simple is the first deliverable of the "Perl Email
Project", a reaction against the complexity and increasing
bugginess of the Mail::* modules. In contrast, Email::* modules
are meant to be simple to use and to maintain, pared to the
bone, fast, minimal in their external dependencies, and
correct.

PR:		56229
Submitted by:	Lars Thegler <lars@thegler.dk>