. Add a port of XawPlus:
XawPlus is a clone of the original Xaw library, providing a more up to date
look and some extensions to the original Athena widget set. This library is
as compatible as possible to the original. It should  be possible to compile
the source code of an Xaw client without any changes (except the include path
of the XawPlus header files) to get the new look of XawPlus to this client.

XawPlus makes it possible to use XPM pixmaps, tooltips, a truncate mechanism
for too long label strings and UTF8 coded UNICODE labels without changing the
code of your application.

XawPlus comes with a set of adapted applications using also the new
features of XawPlus.

WWW: http://people.freenet.de/kra/XawPlus

PR:		47700
Submitted by:	Normal User <freebsd@falx.port5.com>