Add glbsp 2.05,
glBSP is a node builder specially designed to be used with
OpenGL ports of the DOOM game engine. It adheres to the
"GL-Friendly Nodes" specification, which means it adds some new
special nodes to a WAD file that makes it very easy (and fast
!) for an OpenGL DOOM engine to compute the polygons needed for
drawing the levels. There are many DOOM ports that understand
the GL Nodes which glBSP creates, including: EDGE, the Doomsday
engine (JDOOM), Doom3D, PrBoom, and Vavoom.

PR:		ports/61764
Submitted by:	Igor Pokrovsky <tiamat@comset.net>