Add mod_rpaf 0.5
rpaf changes the remote address of the client on incoming proxy
requests. It is used in backend servers that needs to know the IP of the
client, and works in conjunction with mod_proxy_add_forward (which sets
the X-Forwarded-For header). When an X-Host header exists rpaf will take
the hostname found here, put it into Apache, and update the virtualhost
settings so that you can serve virtualhosts on the backend.

Author: Thomas Eibner