Add tcpwatch 1.2.1,
TCPWatch is a utility written in Python that lets you monitor
forwarded TCP connections or HTTP proxy connections. It
displays the sessions in a window with a history of past
connections. It is useful for developing and debugging protocol
implementations and web services.

PR:		ports/66673
Submitted by:	ports@c0decafe.net