- Update to 0.92.8
This update fixes among many other tiny bugs one security bug that
allows a remote attacker to cause a denial of service in bogofilter,
by crashing it; a malformatted (non-conformant) RFC-2047 encoded
word triggers an attempt to write a terminating NUL byte past the
end of a buffer or (more commonly) into the zero-page, which causes
a segfault.

Depending on the exact MTA/MDA configuration on the receiving
machine, this can cause a denial of service of the mail system.

PR:		ports/73144
Submitted by:	maintainer