Update to 0.91.0, the latest version from the window maker team. It contains several bug fixes, especially for the new alt-tab code.
1. Adopt this port, at least for the time being. I use it every day,
and follow the wm-user list so I'm probably qualified.

2. Update the descriptions of the OPTIONS.

3. Restore the knob for KDE, and add knobs for XINERAMA and the
additional LINGUAS files. These, and the XKB_STATUS knobs are all
off by default, but on for the PACKAGE_BUILDING case. I borrowed
this concept from the fluxbox port.

4. Add a knob for the experimental virtual desktop code, off by default.
The window maker folks warn that this is still experimental, so users
shouldn't wander into it by mistake.

5. Instead of bumping the SHLIB version number to 5, set it back to 4.
The window maker folks use N.N.N for their libwraster version string,
and their version for the 0.80.2 version of window maker was 2.3.0.
Thus, our previous version number of 4 was already bogus, so keep it
bogus even though they've bumped their version to 3.N.N. The new lib
is forward compatible for apps compiled with old library (I've tested
many of them), and this "fixes" all the ports that depend on libwraster.

6. Make the locale files hidden under the LINGUAS knob conditional
in the plist.