Upgrade to binutils 2.15.
Revert from USE_LIBTOOL to binutil's own libtool, as libtool13 could
not grok the --without-pic option, and the configure script is not
prepared to handle libtool15 (yet) which in turn could have understood
--without-pic.  Sigh.  At least, that way we can avoid the overhead
of building all the unneeded shared libs.

Still, libtool insists on installing libbfd.la.  As this file is really
small, I give in, and added it to pkg-plist.