- add p5-CPANPLUS-0.53
CPAN++ (also referred to and pronounced as CPANPLUS) is a new and flexible
method of perl module management and installation using the Comprehensive
Perl Archive Network (CPAN). It aims to be a rewrite, and in time a
replacement to the current CPAN.pm. In addition to fixing some long-standing
problems, CPAN++ includes new features, such as module uninstall.

Note: CPANPLUS will become a part of core in perl 5.10

PR:		77012
Submitted by:	Alex Kapranoff <kappa@rambler-co.ru>