Update to 1.1.
Please keep in mind that new versions of this thing just install a new .xpi file
in ${DATADIR}. To upgrade, you must uninstall the old version (via the extensions
manager thinger), restart thunderbird, install the new version, and then restart
thunderbird again.

This info would be useful in /usr/ports/UPDATING, but I'm hesitant to put it
in there for every release.