Add colordiff 1.0.5, a tool to colorize diff output.
Colordiff is a wrapper for diff and produces the same output as diff but with
coloured syntax highlighting at the command line to improve readability.

The output is similar to how a diff-generated patch might appear in Vim or Emacs
with the appropriate syntax highlighting options enabled.

PR:		ports/83062
Submitted by:	Emanuel Haupt <ehaupt@critical.ch>