Upgrade to a new source tarball that has all of our functional private patches integrated.
I give up on trying to convince the doxygen setup of simulavr to
peacefully collaborate with LaTeX 3.x, it always ends up in producing
a PDF file where it is supposed to produce a DVI file (which the build
system then tries to process further).  Instead, I hacked the
configure and build system to not try building or installing any of
its private documentation at all (the man and info pages are still
around as they are already present in the source tarball).

After all, the original simulavr is no longer being developed but has
a successor named simulavrxx, so in future, I'll rather concentrate on
making a FreeBSD port out of that one.