For over five years this port was depending on audio/nas without actually using it, because the nas' libaudio was not linked right.
Now that nas is built better, this port's configure find -laudio to be
usable and builds/installs the `nasay' executable (in addition to the
regular `say').

Put `bin/nasay' back into pkg-plist and changes the bogus
build-dependency on audiolib.h to lib-dependency on audio.

While here, remove the explicit requirement for a particular version of
-lgdbm. Rsynth' last release was in 1994 -- whatever libgdbm happens to
be on a FreeBSD system _today_ is going to be just fine.

Bump PORTREVISION.

Notified by:	pointyhat	(via kris)