Documentation update. Add a blurb at the beginning stating what the file's used for, and add some more comprehensive descriptions to the definable variables.
Also, commit a major faux pas and include a functionality change
along with the text string changes. Remove the conditional check
for adding %D/share/idl to the plist. The check for whether it's
necessary wasn't working, but it's a safe assumption that every
mozilla port needs that dir added to the plist. This change has
been brought to you by ahze and the number n.