Update to 2.6.
This brings back PNG support (for those who missed it with doxygen).

Previously it was not included because of the way it's handled by the
configure script if you use an external version of gd.

This version of graphviz depends upon v2.0.34 of gd, but this version
isn't available from the vendor site. Therefore graphviz uses the
included version of gd. I haven't removed the dependency upon gd, to
let it pickup the port version of gd when it's available. In this case
someone should check if PNG support is still included.

This port disables all new language bindings (like perl, python, ...)
of the 2.6 version to not increase the number of dependencies. Interested
parties should create new ports for those bindings.

Tested with:	devel/doxygen