Add geany.
Geany is a small and lightweight integrated development environment. It was
developed to provide a small and fast IDE, which has only a few dependencies
on other packages. Another goal was to be as independent as possible from a
special Desktop Environment like KDE or GNOME.

WWW: http://geany.uvena.de/

PR:		ports/91817
Submitted by:	Remington <mrl0lz@gmail.com>