Update to 2.4.2alpha after repocopy.
Please note that this is not connected to build at this time, as
the port infrastructure lacks some necessary support for this
OpenLDAP release which will be tested on pointyhat soon.

If you really want to try this out please apply the patch found
in ports/97515 or
  http://www.delphij.net/patch-bsd.port.mk
in order to get infrastructure support for it.

This work is heavily based on ports/96165 submitted by
Joerg Pulz <Joerg Pulz frm2 tum de>.

Approved by:	sem (mentor, implicit)