Follow the new world order: use 'USE_RC_ORDER=file.sh' feature. 	- repocopy vmware-guestd.sh to vmware-guestd.sh.in (by portmgr) 	- remove vmware-guestd.sh 	- remove vmware-guestd.sh entry from pkg-plist 	- update Makefile
It is intensional that I do NOT bump PORTREVISION by this commit, since
any executable is not changed, and startup script itself is also NOT
changed, only the filename is changed.  I do not want to bother users
to update their port by this non-functional change.

If you have any reasonable reason to bump PORTREVISION, let me know, and
I'll update the version asap.