Fix rawio for sequential read/write tests:
Revision 1.3 of patch-ad worked around the problem, that only one writer
is allowed to allow a partition through GEOM. The fix was not complete,
leading to the file position not been incremented during reads and writes,
thus not testing sequential performance, but performance of cached reads
and writes, in general.

This fix makes rawio report reasonable sequential performance again,
but I'm still very suspicious with regard to randomized start positions
working. The results do not show the expected variation of sequential
read/write performance. I have not had time to look into this any deeper,
though, and thus decidied to not delay the commit any further ...