Introducing VMware Tools for VMware Workstation 6 beta 3.
As usual, all three ports are updated after repocoping from
version 5 ports (this is why this ports is 'committed,' not 'imported.')

Note to 7-current users: vmware-guestd and other binaries requires
libc.so.6 to work.  Since there is no misc/compat6x yet, this ports
is marked as 'IGNORE' if there is no libc.so.6.