Add a patch, which disables a call to atexit() to register libjasper's own clean-up routine (jas_cleanup). The call would be of limited use anyway, as freeing memory at exit is useful only for tracking down memory leaks. Removing the atexit call eliminates crashes in ImageMagick and GraphicsMagick, when they are compiled with modules support -- when a library is dlclosed, calling its cleanup routine is certain death...
When compiling with gcc, declare the routine with ``__attribute__
(destructor)'' as per kan's otherwise obnoxious and inflammatory
e-mails. This will make sure, the routine is invoked, when libjasper
is dlclosed(). The only known apps that do that are ImageMagick and
GraphicsMagick (when built with modules support). They both call
the routine explicitly anyway...

While here enable parallel build of jasper itself, and eliminate the
most threatening warnings.

Bump PORTREVISION.