Mark as broken on 5.x: fails to find pthread.h.
Something in the infrastructure changed in the late July timeframe that
actually caused this problem.  The only major thing at that time was the
autoconf/libtool change, but I can't see how that could have caused this
failure mode.  It only happens on 5.x; 6.x and 7.x are fine.

Approved by:	portmgr (self)