Add port math/octave-forge-data-smoothing
This is data-smoothing.
 Algorithms for smoothing noisy data

The octave-forge package is the result of The GNU Octave Repositry project,
which is intended to be a central location for custom scripts, functions and
extensions for GNU Octave. contains the source for all the functions plus
build and install scripts.

PR:             127030
Submitted by:   Stephen Montgomery-Smith <stephen@math.missouri.edu>