Upgrade to the official 6.2 release.
Remove the option to install DRMAA .jar files as they are not
automatically downloadable due to the move the the sun donwload
center.  Adding a new port to install them or adding an option to use
the manually downloaded would probably be a good idea, but I've been
stalling the upgrade too long as it is.