buzhug is a fast, pure-Python database engine, using a syntax that Python programmers should find very intuitive.
The data is stored and accessed on disk (it is not an in-memory database);
the implementation has been designed to make all operations, and especially
selection, as fast as possible with an interpreted language.

The database is implemented as a Python iterator, yielding objects whose
attributes are the fields defined when the base is created ; therefore,
requests can be expressed as list comprehensions or generator expressions,
instead of SQL queries.

WWW:	http://buzhug.sourceforge.net/

PR:		ports/132278
Submitted by:	Wen Heping <wenheping at gmail.com>