Add bugfix for ei_decode_big.
"The high order bits of the 16-bits digit (short) are read, even if
the byte (at u[(i*2)+1]) doesn't belong to the big. In other words,
big numbers such as 16#1234567890 with an odd number of bytes are
decoded with a garbage byte (i.e. 16#XX1234567890)."

http://erlang.org/pipermail/erlang-bugs/2008-October/001023.html

Obtained from:	Paul Guyot <pguyot at kallisys.net>