Upgrade to slony1 version 2.0.2 after repocopy.
From the slony1 home page:

"This is a major new release of Slony-I; it makes use of some
features introduced in PostgreSQL 8.3, and hence is not compatible
with versions older than 8.3. This is considered a pretty good
tradeoff, as various functionality would not be possible with earlier
versions of PostgreSQL."

Support for PostgreSQL 8.4 will be available in slony1 2.0.3, hence
IGNORE_WITH_PGSQL also contains 84 for now.

This port conflicts with slony1-*.

PR:		ports/135096 (with a few changes)
Submitted by:	Ismail Yenigul <ismail.yenigul at endersys.com.tr>
Repocopy by:	marcus