Add a slave port to bind94 for the GeoIP patch from Caraytech.
The most popular use of this patch is to send web site visitors to their
nearest web server. Suppose you have a site called www.example555.com with
two web servers: one in the US and one in England. You can use this patch
in order for visitors from Europe to connect to the server in England and
all other visitors to the server in the US. This is just one example of
its usage. There are probably many others.

WWW: http://www.caraytech.com/geodns/

I created a slave port rather than making this an option but other than
that I was able to use the excellent work in the PR.

PR:		ports/119997
Submitted by:	Jui-Nan Lin <jnlin@csie.nctu.edu.tw>