Akismet is a web service for recognising spam comments. It promises to be almost 100% effective at catching comment spam. They say that currently 81% of all comments submitted to them are spam.
It's designed to work with the Wordpress Blog Tool, but
it's not restricted to that - so this is a Python
interface to the Akismet API.

You'll need a Wordpress Key to use it. This script will
allow you to plug akismet into any CGI script or web application,
and there are full docs in the code. It's extremely easy to use,
because the folks at akismet have implemented a nice and
straightforward REST API.

WWW: http://www.voidspace.org.uk/python/akismet_python.html

PR:		ports/140094
Submitted by:	Jacob Myers (jacob at whotookspaz.org)