Turns out, BLT had its own copies of some important structs, that Tk does not normally export to extensions. Unfortunately, these became slightly different in Tk-8.5 resulting in mistery crashes...
Fortunately, some of these structs are available on FreeBSD, where
we install some important headers, that Tcl/Tk themselves don't
install.  Use those, wherever possible.

Where it is not possible (Frame is defined in BLT's own tkFrame.c),
rename the struct (Frame becomes bltFrame) so that the discrepancy
is easier to catch in the future.

Bump PORTREVISION again.

Tested with: tkdesk, ttt