Update from 1.2.1 to 1.2.2.
Feature safe:	yes

<ChangeLog>

redis-1.2.2 is a bugfix release, fixing a problem with BGSAVE and BGREWRITEAOF
copy-on-write semantics. This version of Redis will use a lot less CPU and
memory while performing a BGSAVE and BGREWRITEAOF operation.

It is very important to upgrade ASAP to this version if you are using Redis
with data persistence via snapshotting or append only file.

Another minor change is that the INFO command will now have a new field showing
memory usage in an human readable form. This was backported from Redis unstable
as well.

</ChangeLog>