Update to 4.72
The changes from the previous release are:

1. TWO SECURITY FIXES: one relating to mail-spools which are
globally writable, the other to locking of MBX folders (not mbox).
These have CVE identifiers CVE-2010-2023 and CVE-2010-2024

2. MySQL stored procedures are now supported.

3. The dkim_domain transport option is now a list, not a single
string, and messages will be signed for each element in the list
(discarding duplicates).

4. The 4.70 release unexpectedly changed the behaviour of dnsdb TXT
lookups in the presence of multiple character strings within
the RR. Prior to 4.70, only the first string would be returned.
The dnsdb lookup now, by default, preserves the pre-4.70
semantics, but also now takes an extended output separator
specification. The separator can be followed by a semicolon, to
concatenate the individual text strings together with no join
character, or by a comma and a second separator character, in
which case the text strings within a TXT record are joined on
that second character. Administrators are reminded that DNS
provides no ordering guarantees between multiple records in an
RRset. For example:

foo.example. IN TXT "a" "b" "c"
foo.example. IN TXT "d" "e" "f"

${lookup dnsdb{>/ txt=foo.example}} -> "a/d"
${lookup dnsdb{>/; txt=foo.example}} -> "def/abc"
${lookup dnsdb{>/,+ txt=foo.example}} -> "a+b+c/d+e+f"

PR:		ports/147686
Submitted by:	Alexey V.Degtyarev <alexey@renatasystems.org>