Add new port pidgin-skype aka. skype4pidgin.
This is a Skype Plugin for Pidgin/libpurple/Adium. It lets you view and
chat with all your Skype buddies from within Pidgin/Adium. You still
need Skype to be running to be able to use it, but it lets you keep a
consistent user interface and use all the other nifty Pidgin/Adium
plugins with it, like spell-checking or OTR encryption. By Eion Robb.