- Add php5-cconv 0.6.4
This port is the PHP wrapper of cconv (chinese/cconv).

cconv (pronunciation: see-conv) is an iconv-based Simplified-Traditional Chinese
conversion tool. It is NOT only a transcoding program, but also a TRANSLATE tool
between Simplified Chinese and Traditional Chinese.

We stand on the shoulders of giants. cconv source code is based on iconv, an
excellent program which converts text from one encoding to another.

Features:
- iconv-like usage.
- language-level Simplified-Traditional Chinese conversion.
- idioms can be translated.

WWW: http://code.google.com/p/cconv/