- update to 4.3.12
Changes since 4.3.11
--------------------

The zsh/parameter module has a new readonly associative array
$usergroups whose keys are the names of system groups of which the
current user is a member and whose values are the corresponding
group identifiers.

The region_highlight array, which controls highlighting of the
command line from zle widgets, is now updated dynamically as
the command line is edited.

In POSIX emulation ("emulate sh") the shell is more accurate about
when it should or should not exit on errors.

The ${NAME:OFFSET:LENGTH} syntax now supports negative LENGTH, which
counts back from the end of the string.

The (g:opts:) flag in parameter expansion processes escape sequences like
the echo and print builtins.  opts can be any combination of o, e and c.
With e, acts like print rather than echo except for octal escapes which
are controlled separately by the o option.  With c, interpret control
sequences like "^X" as bindkey does.  Regardless of the opts, \c is not
interpreted.