This is sieve-connect. A client for the ManageSieve protocol, as specifed in RFC 5804. Historically, this was MANAGESIEVE as implemented by timsieved in Cyrus IMAP.
This is not yet fully compatible with RFC 5804, but is moving towards that from
the timsieved baseline; some issues to be worked on are documented in the
"TODO" file.

sieve-connect speaks ManageSieve and supports TLS for connection privacy and
also authentication if using client certificates. sieve-connect will use SASL
authentication; SASL integrity layers are not supported, use TLS instead.
GSSAPI-based authentication should generally work, provided that client and
server can use a common underlaying protocol. If it doesn't work for you,
please report the issue.

sieve-connect is designed to be both a tool which can be invoked from scripts
and also a decent interactive client. It should also be a drop-in replacement
for "sieveshell", as supplied with Cyrus IMAP.

WWW: http://people.spodhuis.org/phil.pennock/software/

PR:		ports/157077
Submitted by:	Alexey V. Degtyarev <alexey at renatasystems.org>