- New port mail/xpi-dispmua
DispMUA adds an image on the right side of the "header view" which
displays the icon of the application with which the selected mail
was written. So if the person who sent the mail used e.g. mozilla
to send the mail, a mozilla icon will be displayed.

WWW: http://www.juergen-ernst.de/addons/dispmua.html

PR:		ports/149987
Submitted by:	Lapo Luchini <lapo@lapo.it>
Approved by:	maho (mentor)