Fix the build with what seems to have been the intended change from 1.28
It seems that the intention was to make a new BUILD_DEPENDS with the same
content as the existing RUN_DEPENDS, but one line was missed, causing
neither to work. So instead, make BUILD_DEPENDS equal to the old
RUN_DEPENDS, and use the := method to set RUN_DEPENDS, in part because it
helps avoid this kind of error.