- Add p5-SOAP-Transport-FTP 0.711
The SOAP::Transport::FTP module is automatically loaded by the SOAP::Transport
portion of the client structure. It is brought in when an endpoint is specified
via the proxy method that starts with the characters, ftp://. This module
provides only a client class.

WWW: http://search.cpan.org/dist/SOAP-Transport-FTP/