Remove the lang/gcc42 port (GCC 4.2) which does not have any user in the tree any more and has been obsoleted by several newer versions of GCC -- ones that actually build on FreeBSD 10.x out of the box and provide better support for newer CPUs, better optimizations overall, and more advanced language support.
GCC 4.6 aka lang/gcc or lang/gcc46 is the new default version of GCC
in ports.

The system compiler version of GCC 4.2 is still available up to and
including FreeBSD 9.x and users of USE_GCC=4.2+ will be transparently
redirected to lang/gcc if, or rather when, this changes.

PR:		163103
Feature safe:	yes