- Add p5-Module-Implementation 0.03
Module::Implementation abstracts out the process of choosing one of several
underlying implementations for a module. This can be used to provide XS and pure
Perl implementations of a module, or it could be used to load an implementation
for a given OS or any other case of needing to provide multiple implementations.

Module::Implementation is only useful when you know all the implementations
ahead of time. If you want to load arbitrary implementations then you probably
want something like a plugin system, not this module.

WWW: http://search.cpan.org/dist/Module-Implementation/