Reassign to ports@ and deprecate by request of maintainer.
Also, portmgr has been notified that the authors no longer wish to allow
these files to be distributed in a modified fashion, as they are intended
to be used for conformance testing.  If anyone is interested in working
with the upstream to integrate FreeBSD's patches, please contact portmgr.

Hat:		portmgr