After repocopy (by marcus@) from linux_base-f10:
Add linux_base-c6 port. This is a linux_base port based upon CentOS 6.
CentOS 6 is supported upstream (so far) until November 30, 2020.

Compared to the port as submitted in the PR, I added most SRC_DISTFILES
and set the maintainer to emulation@ (discussed with the submitter).

Not connected to the ports collection (emulation/Makefile) yet, first I want
to find the missing src RPMs (compat-*, stdc++, ...). The port is committed
anyway, so that people can help with the TODO list (in the Makefile).

WARNING: it is highly experimental to use this port instead of the default
linux_base port. You need to set compat.linux.osrelease to 2.6.18, so watch
out for unimplemented syscalls (most recent 8-stable / 9-stable or -current
is recommended for this).

Submitted by:	Svyatoslav Lempert <svyatoslav.lempert@gmail.com>
PR:		165900