- New port sysutils/autojump
Autojump is a tool that acts as a complement to cd: it makes navigating your
filesystem a lot faster. It works by automagically maintaining a database of
the directories you use the most from the command line, and allows you to jump
back and forth between them, by typing just a few letters of the name of the
directory you want to jump to.

PR:		ports/151467
Submitted by:	Neeraj Verma <neeraj.verma.ports@vermatech.com>