New Features ============ Print the same progress update that goes to the term title in-line, and apply various improvements to that code.
Move 'make clean' to the background. There is no point waiting around
for it to finish. In the case of very large work directories this may
cause some I/O contention during the unpacking of the next port to
build, but in most cases this should be lost in the noise.

Other Improvements
==================
Speed up the determination of WRKDIRPREFIX, and do not bother looking
for empty directories if we exited abnormally.

Add code to detect dependency loops created by moving part of a port
into another port, and CONFLICT'ing on previous versions of that port
(ala textproc/p5-XML-SAX -> p5-XML-SAX-Base). When a loop is detected
suggest that the user try pkg_updating.

Bug Fixes
=========
Improve error handling for mktemp

Improve handling of abnormal exits (such as ^C, failed port, etc.),
especially when the first exit happens in a child process.

If the second argument to -o is something that is completely non-existent,
report in the error message what the user typed on the command line.
Reported by:	avilla

In clean_build_only_list() remove ports that are already installed
when using --delete-build-only to prevent them from being spuriously
removed.
Reported by:	avilla