Apply patch to make the port build with libc++.
libc++'s cwchar header introduces an ambiguity with clang that makes a
few different versions of wcschr() be found. Talking to theraven@ it
looks like a problem in the C++ standard itself, so there is not much
more to do other than working around the issue with this patch.

Since the clucene copy in Qt is old and has no maintainer, there is
not much value in committing this upstream, even though it may be
worth a try in the future.

No PORTREVISION bump, as users of libstdc++ are not affected, and no
dependencies have been changed.

PR:               ports/169872
Submitted by:     Jan Beich <jbeich@tormail.org>