- New port math/octave-forge-ncarray.
The octave-forge package is the result of The GNU Octave Repositry project,
which is intended to be a central location for custom scripts, functions and
extensions for GNU Octave. contains the source for all the functions plus
build and install scripts.

This is ncarray.

Access a single or a collection of NetCDF files as a multi-dimensional array.