Add x11-themes/kde4-style-oxygen-transparent.
Oxygen style and decoration with support for transparency.

This is an official branch of the KDE oxygen style. It is kept in sync with the
official oxygen style, and adds support of full transparency on the windows.

The style and the decoration are named "Oxygen Transparent" and appear as such
in KDE's system settings (in both the applications and workspace appearance
pages).

A configuration helper application is included with the style that allows one to
configure both the style and the decoration in the same window. It is called
oxygen-transparent-settings.

WWW: http://kde-look.org/content/show.php?content=127752

PR:		172733
Approved by:	bdrewery (mentor)