Add patches to gamin to enable support for effective uid/gid, add a option for this and turn it on by default.
Current gamin uses the current real uid to determine the location of the
socket and during the security routines. It also forks off a copy of
"gam_server" as the current real uid as well.

This breaks gamin functionality with any application that uses effective an
uid/gid for privilege separation, such as samba. Samba runs as root, but
always sets the effective uid/gid to that of the connecting user. Since
gamin is only aware of the real uid, it always creates a root-owned gamin
socket that the effective user can't touch (gamin's own security checks
refuse it).

PR:		ports/176203
Submitted by:	Danny Warren <danny@dannywarren.com>
Reviewed by:	marcus@