Add emulators/i386-wine-devel, slave to emulators/wine-devel.
This port adds shims to enable wine to run cleanly on an amd64 system.
This is done by doing a "static link" of the package: including all
libraries required by the port.

It is possible to use this package under i386, although it is intended
for an amd64 system.

Feedback welcome for improvement of the port.

Approved by:	bdrewery (mentor)