Fix build when WITH_CCACHE_BUILD is set.
The problem is that WITH_CCACHE_BUILD adds PATH to the MAKE_ENV, but
the gnustep support sources a GNUStep.sh file that already sets the
PATH. Next when env PATH=... is called, it overwrites and losers the
gnustep PATH additions. This is a temporary fix until a better
solution is found.