New port: devel/kyua-atf-compat:
Kyua (pronounced Q.A.) is a testing framework for both developers and
users.  Kyua is different from most other testing frameworks in that it
puts the end user experience before anything else.  There are multiple
reasons for users to run the tests themselves, and Kyua ensures that
they can do so in the most convenient way.

This module, kyua-atf-compat, provides tools to ease the transition from
ATF to Kyua.  In particular, this includes is a tool to convert Atffile
files to Kyuafile files in an automated manner, and drop-in replacement
wrappers for atf-run and atf-report.

WWW: https://code.google.com/p/kyua/

PR:		ports/177640
Submitted by:	asomers
Reviewed by:	Garrett Cooper <yaneurabeya@gmail.com>