x11-toolkits/gtkada: Improve multijob support
The first attempt at supporting multiple jobs resulting in reimplementing
the MAKE_JOBS_NUMBER determining with "!=" operator.  Portlint doesn't
like that and it's not even necessary.  By overriding the do-build target,
the "-j" parameter is shielding from the makefile and the number of jobs
is passed to gprbuild via the PROCESSORS environment variable.

USE_GMAKE was also converted to USES+= gmake.

Approved by:	bapt (mentor)