devel/gps: Add multijob support
This is a big port that takes a lot of time to build, so enabling multijob
support is a big win.  As with other Ada programs, the "-j" configuration
argument cannot be passed to the Makefile.  The multijob support is enabled
through the PROCESSORS environment variable.  These requirement necessitate
a custom do-build target.

Some minor portlint cleanup is done as well.

Approved by:	bapt (mentor)