Add new port: lang/asis
This is ASIS, the Ada Semantic Interface Specification.
ASIS is tightly coupled with lang/gcc-aux and must be built with the same
sources.  For this reason, ASIS will not work with other Ada compilers,
including GNAT GPL (lang/gnat).

ASIS is a dependency of several upcoming Ada ports.

Approved by:	bapt (mentor)