sysutils/gkrelltop: Upgrade to 2.2.10 to unbreak fetch
The MASTER_SITES of the current version of gkrelltop, 2.2.6 has disappeared.
The project moved to SourceForge, but the earliest version available there
is 2.2.7.  However, 2.2.7 failed to build on FreeBSD 8.4 i386 and
FreeBSD 9.1 amd64.  Version 2.2.13, the latest, fails to build on all
platforms, including DragonFly.  Upgrading to version 2.2.10 results in
successful builds on all platforms.

Also, the header was trimmed and USE_GMAKE was replaced with USES+=gmake.

PR:		ports/178534
Approved by:	bapt (mentor), maintainer timeout (~10 weeks)