textproc/words: Move to Ada framework
Currently this port has a hard dependency on lang/gnat.  This has been
replaced with USES+=ada and minor changes to incorporate it into the Ada
framework with a PORTREVISION bump.

The MASTER_SITES url did not resolve, so the distfile was unfetchable.
Luckily the entire site was archived, so the MASTER_SITES and primary
site fetching has been restored.  The makefile header was trimmed, the
AUTHOR was removed as the email may not longer be valid, NOPORTDOCS
was replaced with the DOCS option, and now portlint is happy.

Approved by:	bapt & culot (mentors)