Enable packaging of emulators/i386-wine-devel on amd64 (part 2/2).
CHANGES
-------
Provide two ports, in one.  When compiling on i386 the port behaves
as a slave port of wine-devel, creating a package suitable for
installation on amd64.  No change here

When compiling on amd64 the port manually installs the provided
amd64 packages (see wiki.FreeBSD.org/i386-Wine for those packages)
and thus allowing the packages to be "build" using the FreeBSD
infrastructure, and critically, to appear in the standard package
set without requiring users to manually add these packages to their
systems.

DESIGN
------
The bootstrapping (for choosing between i386 and amd64 Makefiles)
was done manually as Makefile.i386 is a slave port while Makefile.inc
(for amd64) is a master port.  This situation does not work in the
current infrastructure thus requiring a manual bootstrap.

PRECEDENT
---------
Although Ports does not support cross compiling of packages there is
precedent in supplying binary packages for those situations where,
otherwise, cross compiling is required.

In support of taking this approach I site:
 - misc/compat?x ports
 - */linux-* ports