Update emulators/i386-wine-devel options with GECKO and MONO.
The binary package for amd64 systems does not bundle GECKO or MONO
however it is useful (for some) to have those files installed, so
allow the package to have a run-time dependency on the ports that
provide Gecko and Mono support.

PORTREVISION is not bumped since nothing changes in the default
(BATCH) case.