- New port math/octave-forge-netcdf
The octave-forge package is the result of The GNU Octave Repositry project,
which is intended to be a central location for custom scripts, functions and
extensions for GNU Octave. contains the source for all the functions plus
build and install scripts.

This is netcdf: A MATLAB compatible NetCDF interface for Octave.