Add a ton of patches needed to get the port building with libc++.
One of them is a patch I got into Qt some time ago, but patch-libcxx is a
collection of commits to WebKit itself (and a few changes of mine) that
never got downstream Qt's copy of WebKit.

Some good soul with free time should work on porting qtwebkit-23 [1], which
contains a much more up-to-date WebKit checkout.

[1] https://blogs.kde.org/2013/07/17/qtwebkit-232-and-qtwebkit-qt-51

PR:		ports/181932