sysutils/ciso: unbreak, de-deprecate
I mistaken thought Rodrigo was the maintainer of ciso.  He straightened that
out.  Vasiliy, the maintainer, states that ciso in practical terms is still the
latest version, and provided a patch to fix the fetch breakage.  The expiration
and deprecation has been removed.

PR:		181558
Submitted by:	blttll (Vasiliy the maintainer)