net-im/qutim: add <unistd.h> for modern compilers, fix binutils
In addition to the header inclusion, the linking stage failed on modern
binutils because libX11 was used without being explicitly listed.  The
LDFLAGS from the Makefile were getting overwritten by the CMake makefile,
so that had to be patch to allow the LDFLAGS to be appended.

Approved by:	portmgr (bapt, implicit)