cad/varkon: Hardcode PLATFORM to "FreeBSD"
The use of $OPSYS to define the platform actually breaks the build on
DragonFly.  The PLATFORM has to be recognized, and it serves for file
selection.  DragonFly (the only other OPSYS using ports for now) needs
"FreeBSD", not "DragonFly" to be defined as PLATFORM in order to build.
This change has no effect for FreeBSD users.

Varkon is an unmaintained port.