Mk/Uses/compiler.mk: Ensure COMPILER_TYPE can only be clang or gcc
The compiler.mk comments and code state that COMPILER_TYPE can only be
of the value "clang" or "gcc".  However, the code that determines this
allows for a possible undefined third state (empty string).  BMake
will emit a lot of errors about badly formatted conditionals if
COMPILER_TYPE is empty.

Since, by definition, if the COMPILER_TYPE is not clang, it must be
gcc, so skip the conditional gcc check and just set it.  The entire
file must be updated if support for additional compilers is desired.

This bug was discovered because the gcc detection code failed to
identify the DragonFly base compiler (GCC 4.7.3) as gcc.

Approved by:	portmgr (bapt)