editors/xxe: Mark BROKEN and DEPRECATED, expire in 1 month
This unmaintained port is also a high-maintenance port.  Every 3-4
months it becomes unfetchable as the current distfile is removed
and replaced with a new version.  This is intentional on the part
of upstream because it is trialware with restrictions.

In order to keep this port, somebody must actively maintain it,
as it's requirements are not reasonable for ports@

Approved by:	portmgr (bapt)