Prepare devel/netbsd-pkgsrc-mk-files for removal.
This port was added to the tree because devel/mk-configure
depended on it. netbsd-pkgsrc-mk-files is a collection of mk
files from NetBSD pkgsrc project, packed by the author of the
mk-configure. Later he included the necessary files to
mk-configure, removed the dependency and netbsd-pkgsrc-mk-files
is not updated since then.

Approved by:	bdrewery (mentor)
PR:		ports/182724