lang/gnatdroid-binutils: Remove USE_GCC=yes
A pkg-fallout notice was received for gnatdroid, saying that lang/gcc and
lang/gnatdroid-* were installing files in the same place.  It turns out
that gnatdroid-binutils was pulling lang/gcc in unnecessarily.  With
--disable-werror, it builds on clang on F10 just fine.

While here, convert USE_GMAKE too.  Soon all gnat* ports will be
updated with the latest for ports, this is just a stop-gap fix.