Please welcome NSD 4.0.0
For all new features, see
http://www.nlnetlabs.nl/svn/nsd/tags/NSD_4_0_0_REL/doc/NSD-4-features

This version replaces the nsdc control program with nsd-control.
This requires some manual setup with nsd-control-setup and editing
of the config files. nsd-control is incompatible with nsdc so when
that is used in scripts, these should be adapted.

NSD 3 is still supported as dns/nsd3.

PR:		183888
Submitted by:	Jaap Akkerhuis <jaap@NLnetLabs.nl>