Fix build of audio/soundkonverter due to the use of '-pedantic'.
Soundkonverter includes a C file that has a trailing comma (,) at the end
of an enum, this construct is only supported by C++11 (but by C99) thus is
not accepted as valid with the '-pedantic'.  The base compiler (gcc) does
not understand C++11 and errors on this.

Remove the '-pedantic' option as it does not impact on the code produced,
to allow the port to be built.

PR:	ports/182906