Add multimedia/livestreamer:
Livestreamer is a Command Line Interface that pipes video streams
from various services into a video player, such as VLC. The main
purpose of Livestreamer is to allow the user to avoid buggy and CPU
heavy flash plugins but still be able to enjoy various streamed
content.

Livestreamer is built upon a plugin system which allows support for
new services to be easily added. Currently most of the big streaming
services are supported, such as:
    Dailymotion
    Livestream
    Twitch/Justin.tv
    YouTube Live
    UStream

Example use:

$ livestreamer twitch.tv/day9tv best
[cli][info] Found matching plugin justintv for URL twitch.tv/day9tv
[cli][info] Opening stream: 720p
[cli][info] Starting player: vlc

WWW: https://github.com/chrippa/livestreamer