lang/python26: Use ports libffi to address ctypes build failure on i386
Unconditionally use libffi from ports because the Python 2.6 branch was
closed for maintenance when the import of libffi 3.0.13 took place.
fixes _ctypes module build failure on i386. [1]

This change was also committed to lang/python31 [2] and makes the recent
FreeBSD/ARM patch [3] a NOOP, which will once again be addressed when
ports/184517 lands [4]. This gives users the greatest amount of choice
and flexibility just prior to expiry of the python27 and python31 ports

[1] Fixes Issue #17192: Update the ctypes module's libffi to v3.0.13.
    This specifically addresses a stack misalignment issue on x86 and
    issues on some more recent platforms. [#17192]

    http://bugs.python.org/issue17192
    http://hg.python.org/cpython/rev/a94b3b4599f1
    http://hg.python.org/cpython/rev/688bc0b44d96

[2] https://svnweb.freebsd.org/ports?view=revision&revision=335370
[3] https://svnweb.freebsd.org/ports?view=revision&revision=335867

PR:	ports/184517 [4]