The MCPU_OPTS conditional block currently incorrectly sets ABI=64 for all ARCH strings that in 64. GMP expects mode64, mode32 or 32 as valid values for the ABI option [1]. Using an invalid value causes the following build error on powerpc64:
    configure: error: ABI=64 is not among the following valid choices:
    mode64 mode32 32

This is the minimum change to set the correct "mode64" ABI value when
ARCH is powerpc64, while also keeping all other semantics the same. A
more complete refactor of this port is possible, but would need an
exp-run to test for regressions.

The change was tested on powerpc64 (thanks justin) and amd64 (for regression)

While I'm here:

- Remove indefinite article from COMMENT

[1] https://gmplib.org/manual/Build-Options.html

PR:		ports/179127
Submitted by:	jhibbits
Reviewed by:	jhibbits, ak
Approved by:	maintainer timeout (6 months)