editors/beav: Unbreak after staging support attempt
After recent commits, beav was broken on every platform for a couple
of reasons.  Primarily the Makefile had a bug in it, so it was
unexecutable.  Secondly, the "man page fix" was done in ${STAGEDIR}
using sed -i, causing the resultant .bak file to also be installed, or
at least orphaned.  Poudriere caught that issue.

While here, add DragonFly support.