Add x11-fonts/clearsans.
Clear Sans is a versatile OpenType font for screen, print, and Web.

It was designed with on-screen legibility and glanceability in mind. It strikes
a balance between contemporary, professional, and stylish expression and
thoroughly functional purpose. It has a sophisticated and elegant personality
at all sizes, and its thoughtful design becomes even more evident at the thin
weight.

WWW: https://01.org/clear-sans