Go back to 6.8.0 :(
This reverts r338918, r338948, r338951 and r338953 (modulo reassigning the
port to ports@, some comment fixes and the removal of a patch that was meant
for FreeBSD 7).

ImageMagick is normally a beast to update, and somewhere between 6.8.0 and
6.8.8 the developers also decided to change the naming scheme for their
libraries, which requires updating tons of ports in the tree and bumping a
lot of PORTREVISIONs. I do not have the time for that at the moment, so it's
safer to go back to what was working and let a braver soul handle the update
properly later, otherwise the tree will be on fire for quite a while.

ache: I'm truly sorry for reverting your PerlMagick commits, but I couldn't
test if they still worked with the older version in a timely fashion.