lang/gcc-aux: Update diff-ada, diff-core for Android
The majority of these changes affect the android cross-compiler builds.
The only build change for the host compiler is builds convert_addresses
within a separate file rather than as part of adaint.c.  Given that this
does not represent a functional change in any way, I just don't see the
need to bump PORTREVISION as a result.