Welcome plexmediaserver-plexpass to ports
This is a clone of the existing multimedia/plexmediaserver port.

All future "Plexpass" (read: bleeding edge) releases will land here on
this port. Subsequently, all future public/stable releases of Plex Media
Server will be committed to the existing multimedia/plexmediaserver
port.

Now that Plex has stabilized on FreeBSD and users are not waiting for a
laundry list of platform-specific bugfixes it is important that we offer
choice of both the recommended stable version as well as the bleeding edge.