devel/smake: Unbreak smake to see what happens
This port was originally marked as broken on i386 only.  It always worked
without modification on dports.  Later when eadler@ released the port
without fixing i386, he marked it completely broken although it builds on
amd64.  He did not post the error messages encountered on i386 in the
commit message adding BROKEN_i386 or BROKEN and thus all smake error logs
have cycled off.

Remove BROKEN definition and let the cluster start building smake again
so we can see if it's really broken or not (and if so, how).