Fix problematic package upgrade
Unifi comes with a WAR file that it unpacks if the webroot does not exist.
This was causing the webroot to be untracked. Upon an upgrade Unifi
would not clean up the old webroot or unpack the new WAR file leaving
the web interface non-functional. This is solved by unpacking the WAR
file before creating the package so we track these files.