Uses/ada.mk: Add support for soon-to-be-older Ada compiler
Currently the Ada framework uses a single compiler based on gcc47.
Soon this compiler will be updated to the gcc49 base, but not all Ada
ports are currently happy about this.

This update to ada.mk induces the ARGS of "47" which causes the port to
use the newly-created lang/gcc47-aux port which is currently a clone
of lang/gcc-aux.  When the latter is updated, the Ada Framework will
use it except for those ports that feature "USES+= ada:47"