devel/florist-gpl: Fix code style before gcc49 comes in
Part of florist is built with -gnatg which invokes style checking.
The upcoming GNAT compiler has stricter rules regarding the placement
of the "then" keyword.  Florist was failing to build based on style
errors on gcc49.  Patch this now on gcc47 to prepare for the Ada
compiler transition.

No bump; the code logic doesn't change.