lang/gcc-aux: Upgrade from gcc-4.7.3 to gcc49-20140302 snapshot
GCC4.9 hasn't been released yet, but the Ada portion of it is pretty
stable and perfectly passes ACATS and gnat.dg testsuites.  The only
trick was to change the unwind mechanism on i386 away from KERN_PS_STRINGS
because it will not work with 32-bit compatability mode on AMD64.  This
is because on AMD64, the signal trampoline is located on a random area
on a shared page rather than a fixed location before KERN_PS_STRINGS.  To
support both true i386 and AMD64-32 compatibility mode, unwind now looks
for frame pattern in all cases.  Apparently stack checking has been broken
on AMD64-32 until now.

It is important to get the Ada Framework in ports to be based on gcc49
because this compiler supports the full Ada-2012 standard implementation,
and in fact some new programs require this already.

This port will stay on snapshots until 4.9.0 is released, then it will
return to be based on releases.

The Ada and C++ options have been removed.  These are now always built.
In fact, GCC now requires C++ to build, so new bootstraps had to be
created.  FreeBSD8 has its own bootstrap, and FreeBSD 9+ uses a pure
static bootstrap (including GNAT tools) which should last a long time.
Additionally a "Bootstrap" option was added, but this overrides all
other options to create new bootstraps. It is not intended for general
use.

Most of the Ada Framework in ports has been pretested and adjusted to
work with gcc49 out of the box, but some ports will need to be updated
immediately, which will happen right after this one.  The update to
lang/gnat_util had to be synchronized with lang/gcc-aux because they
share version information and the PORTREVISION needed to be reset.