lang/mono: Unbreak configure/runaway
Mono was set to build via clang unconditionally.  Other than have
a large, unnecessary dependency on clang33, this didn't bother
DragonFly until recently.  The 3.2.8 version has a conftest to test
for a working __thread implementation (which DragonFly has), but for
some reason the test locks up when built by clang.  I built the conftest
manually with gcc and it returned fine.

In any case, DragonFly's gcc47 base compiler builds mono fine, so to
unbreak the port on DragonFly, only specify clang on FreeBSD.