www/seamonkey: Unbreak DragonFly
The first fix is a NO-OP for FreeBSD.  The clang/optimized flags work-
around is OSVERSION-specific, so let's limit it to FreeBSD only.

The real breakage was caused by the XPI_FILE substitution.  The "x86"
part of DragonFly's "x86_64" was getting transformed incorrectly to
"i386_64".  This could be prevented by changing the replacement pattern
to "x86$", but "x86" is not a standard ARCH value, so this substitution
isn't made on FreeBSD anyway.  The solution is just remove it as the
only effect it has is negative.

Tested on: Redports, all four i386 builders pass