textproc/xmlada, www/aws: Hack to fix build on FreeBSD 8
For an unknown reason, the shared version of the xmlada library fails
to link on FreeBSD 8.  To at least get it building, only the static
version is built on that platform.  As a consequence, www/aws had to
have a similar FreeBSD8-only instruction.  Neither requires plist
adjustments because they both have generated plists.  No bump is
required because neither have been building on FreeBSD 8.

xmlada is due to be updated anyway.  Most Ada libraries are static and
I very well just convert xmlada to be a static-only library without
condition.

While here, fix a check-orphans failure due to licensing on aws.