lang/gcc-aux: Upgrade from 30 March snapshot to 6 April
Since the April 6 snapshot, a lot of work has entered the GCC tree to
fix the arm-android GNAT port.  Currently it passes everything except
stack checking.

A lot of the additional lines in diff-ada will be removed with the
next snapshot.  Future plans also include significantly reducing the
size of the diff-cxx-testsuite file too.