The PWCBSD option requires V4L, so remove -DWITHOUT_V4L.
I'm not sure why this change was made, but it breaks the PWCBSD
option. It looked obviously wrong; it added a dependency on v4l_compat
and then disabled it by definining WITHOUT_V4L. Looking at the code
it's obvious V4L is required for anything other than network cameras
too.