- Fix the plist when the CFITSIO option is on. - Use the new OPTIONS helpers to simplify the Makefile.
Related to both, also stop automatically turning the CFITSIO option on when
libcfitsio.so is installed: wcslib has had proper --with{out}-cfitsio
options since version 4.4.4.

This will turn off CFITSIO support for most people and indirectly fix their
builds with astro/cfitsio-3.360-1 [1]. If people do choose to have it on,
ports/189487 needs to be committed, otherwise there will be plist errors
because wcslib's configure script will not be able to detect that cfitsio is
installed and can be used.

[1] http://lists.freebsd.org/pipermail/freebsd-ports/2014-May/092070.html