devel/binutils: Support dev branch of DragonFly
On the development branch of DragonFly, a number of shared system
libraries has moved from /usr/lib to /lib, mirroring what happened on
FreeBSD years ago.  The bfd linker will not follow symlinks, so the
configure.tgt file needs to be updated for DragonFly 3.7+.  This is a
no-op for FreeBSD.

A similar patch has been submitted upstream.