devel/gps: Upgrade version 5.2.1 => 6.1.0.0
Adacore releases the GNAT Programming Studio annually along with
several other packages.  It is under constant development, so the source
tarballs contain specific versions of xmlada, gtkada, and all the other
dependencies.  However those dependencies are under development as well
and the current versions in the port collections were too new for the
actual 2014 GPS release (which was frozen 7 months ago).

So I packaged the working version of the upcoming 6.1.0 version instead
(the 4th digit means "working" and it increments each time the distfile
is upgraded).

Several options were removed because they are expected by GPS components
to be present (e.g. sqlite and gtk).  The Iconv support was intended to
be optional, but turning it on crashes GPS after the splash popup.  Until
the issue is resolved, the Iconv option has been removed and the GPS
has been configured without it.

GPS 6.x differences from the 5.x series in that the GTK3 toolkit is used
instead of GTK2.  It also has language support for SPARK 2014, syntax
highlighting, tooltips for Ada 2012 and SPARK 2014, enhancements to the
scripting API and various editor enhancements.

Release Notes with lots of pictures illustrate new capabilities here:
http://docs.adacore.com/gps-docs/release_notes/build/singlehtml